+++ internal/component/kafka/sasl_oauthbearer.go
