+++ tests/config/bindings/kafka/bindings.yml
