+++ tests/config/pubsub/kafka/pubsub.yml
@@ -14,5 +14,3 @@ spec:
     value: "false"
   - name: initialOffset
     value: oldest
-  - name: disableTls
-    value: true
